+++ src/app/login/login.component.ts
@@ -33,7 +33,7 @@ export class LoginComponent implements OnInit {
           this.authService.logIn(this.memberID, this.pass);
           this.router.navigateByUrl('/account');
         }
-        this.loginService.logIn(this.username, this.pass, response.body.memberID);
+        this.authService.logIn(this.username, this.pass, response.body.memberID);
         this.router.navigateByUrl('/account');
       },
       (error) => {
